Small refactoring

+++ TheFall/src/DormanPrinceSolver.cpp
@@ -100,9 +100,9 @@ void DormanPrinceSolver::Run(Model &model)
 		sum.fill(0.0);
 		sum_1.fill(0.0);
 
-		for (int i = 0; i < x_size; i++) // проходим по элементам вектора Х
+		for (size_t i = 0; i < x_size; i++) // проходим по элементам вектора Х
 		{
-			for (int j = 0; j < SIZE; j++) // собираем воедино все k
+			for (size_t j = 0; j < SIZE; j++) // собираем воедино все k
 			{
 				sum(i) += m_b(j) * m_k(j, i);
 				sum_1(i) += m_b1(j) * m_k(j, i);
